Extract workspace form state helpers
